Rework meson.build; Switch to Linux style
The documentation and "general wisdom" of how to use Meson seems
drastically incorrect, which caused much of my prior confusion. You
can in fact dynamically control build options based on buildtype in a
way that seems pretty clean compared to what I've seen in the wild so
to speak. I also moved the entire build details into a single file
which should make it easier to see exactly what's happening, like the
addition of fsanitize for debug builds and lto + strip for release
builds.

The AddressSanitizer is unhappy about some of the code, and I'm not
happy with most of it (even for a toy project). I'll enable it again
once I've worked through whether it's errors/regressions (not likely)
or terrible coding semantics (very likely).

https://reviews.llvm.org/rGf787cf7ee65018d99a9da6676e76079a97847bf8