Switch to OpenBSD style(9), or at least as close as I can get
After reading the Clang Format options a few times I think this is
realistically as close as it can get to OpenBSD KNF/style(9) semantics
without work on clang format itself. The two biggest issues are not
being able to control break semantics on macros, and right-aligned
pointers have a long-standing style bug which introduces padding.

It doesn't have to be perfect, but in most other respects it appears
to be quite spot on.